Move test logger helper to new module for reuse
